refactor(metrics): move rtc device metrics out of logger
Move rtc device metrics out of logger to devices.
Flush rtc device metrics through legacy flush_metrics().
