+++ src/vmm/src/devices/legacy/mod.rs
@@ -8,7 +8,7 @@
 //! Implements legacy devices (UART, RTC etc).
 mod i8042;
 #[cfg(target_arch = "aarch64")]
-mod rtc_pl031;
+pub mod rtc_pl031;
 pub mod serial;
 
 use std::io;
@@ -68,5 +68,7 @@ impl EventFdTrigger {
 pub fn flush_metrics<S: Serializer>(serializer: S) -> Result<S::Ok, S::Error> {
     let mut seq = serializer.serialize_map(Some(1))?;
     seq.serialize_entry("i8042", &i8042::METRICS)?;
+    #[cfg(target_arch = "aarch64")]
+    seq.serialize_entry("rtc", &rtc_pl031::METRICS)?;
     seq.end()
 }
